moved function evaluate to package excel, created function for it
